\name{ppks-package}
\alias{ppks-package}
\docType{package}
\title{
Permutation Based Paired Kolmogorov-Smirnov Test 
}

\description{
Permutation Based Paired Kolmogorov-Smirnov Test 
}

\details{
\tabular{ll}{
Package: \tab ppks\cr
Type: \tab Package\cr
Version: \tab 1.0 \cr
Date: \tab 2025-09-21\cr
License: \tab GPL-2\cr
}
}

\author{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\section{Maintainers}{
Michail Tsagris <mtsagris@uoc.gr>
}

\references{
Wang W.S., Amsler C. and Schmidt, P. (2025). A randomly swapped bootstrap for paired data: testing equality of distribution for correlated samples. 
Empirical Economics, To appear.
https://link.springer.com/article/10.1007/s00181-025-02779-0
}

%\note{
%}


